+++ app/partials/attachment.php
@@ -2,10 +2,10 @@
 declare (strict_types = 1); ?>
 
 <div class="entry-attachment">
-    <p><a href="<?php
-        echo \wp_get_attachment_url($post->ID);
-    ?>" rel="attachment" itemprop="url"><?php
-        echo \basename($post->guid);
+    <p><a href="<?=
+        \wp_get_attachment_url($post->ID);
+    ?>" rel="attachment" itemprop="url"><?=
+        \basename($post->guid);
     ?></a></p>
 
     <?php if ($post->post_excerpt) { ?>