+++ app/templates/index.php
@@ -11,8 +11,8 @@ if (!\Jentil()->utilities->page->is('singular')) {
         <header class="page-header">
             <?php \do_action('jentil_before_title'); ?>
 
-            <h1 class="page-title" itemprop="name mainEntityOfPage"><?php
-                echo $jentil_title;
+            <h1 class="page-title" itemprop="name mainEntityOfPage"><?=
+                $jentil_title;
             ?></h1>
 
             <?php \do_action('jentil_after_title'); ?>